Add degree functionality to subject controller

+++ Backend/src/controllers/subject.controller.js
@@ -2,20 +2,50 @@ import { Subject } from "../models/subject.model.js";
 import { ApiResponse } from "../utils/ApiResponse.util.js";
 import { ApiError } from "../utils/ApiError.util.js";
 import { asyncHandler } from "../utils/asyncHandler.util.js";
+import { Degree } from "../models/degree.model.js";
 
 const createSubject = asyncHandler(async (req, res) => {
-    const { title, years } = req.body;
+    const { title, years, degreeId, degreeTitle } = req.body;
 
-    const yearsArray = years.split(' ').map(value => Number(value));
+    const yearsArray = years
+        .split(" ")
+        .map((value) => Number(value))
+        .sort();
 
     if (!title || !years || !Array.isArray(yearsArray)) {
         throw new ApiError(400, "Title and years are required");
+    }
+
+    if (!degreeId && !degreeTitle) {
+        throw new ApiError(400, "Degree id or title is required");
+    }
+
+    const degree = await Degree.findOne({
+        $or: [{ _id: degreeId }, { title: degreeTitle }],
+    });
+
+    if (!degree) {
+        throw new ApiError(404, "Degree not found");
     }
 
     const subject = await Subject.create({
         title,
         years: yearsArray,
     });
+
+    if (!subject) {
+        throw new ApiError(500, "Something went wrong while creating subject");
+    }
+
+    degree.subjects.push(subject._id);
+    try {
+        await degree.save();
+    } catch (error) {
+        throw new ApiError(
+            500,
+            "Something went wrong while saving degree" + error.message,
+        );
+    }
 
     return res
         .status(201)
@@ -46,7 +76,10 @@ const updateSubject = asyncHandler(async (req, res) => {
     const { id } = req.params;
     const { title, years } = req.body;
 
-    const yearsArray = years.split(' ').map(value => Number(value));
+    const yearsArray = years
+        .split(" ")
+        .map((value) => Number(value))
+        .sort();
 
     if (!title && !(years && Array.isArray(yearsArray))) {
         throw new ApiError(400, "Provide at least one field to update");
@@ -60,7 +93,7 @@ const updateSubject = asyncHandler(async (req, res) => {
                 ...(yearsArray.length && { years: yearsArray }),
             },
         },
-        { new: true }
+        { new: true },
     );
 
     if (!subject) {
@@ -74,10 +107,34 @@ const updateSubject = asyncHandler(async (req, res) => {
 
 const deleteSubject = asyncHandler(async (req, res) => {
     const { id } = req.params;
+    const { degreeId, degreeTitle } = req.body;
+
+    if (!degreeId && !degreeTitle) {
+        throw new ApiError(400, "Degree id or title is required");
+    }
+
+    const degree = await Degree.findOne({
+        $or: [{ _id: degreeId }, { title: degreeTitle }],
+    });
+
+    if (!degree) {
+        throw new ApiError(404, "Degree not found");
+    }
 
     const subject = await Subject.findByIdAndDelete(id);
     if (!subject) {
         throw new ApiError(404, "Subject not found");
+    }
+
+    degree.subjects = degree.subjects.filter((s) => s._id.toString() != id);
+
+    try {
+        await degree.save();
+    } catch (error) {
+        throw new ApiError(
+            500,
+            "Something went wrong while saving degree" + error.message,
+        );
     }
 
     return res
@@ -85,4 +142,10 @@ const deleteSubject = asyncHandler(async (req, res) => {
         .json(new ApiResponse(200, subject, "Subject deleted successfully"));
 });
 
-export { createSubject, getSubjects, getSubjectById, updateSubject, deleteSubject };
+export {
+    createSubject,
+    getSubjects,
+    getSubjectById,
+    updateSubject,
+    deleteSubject,
+};
